+++ benchmark/project/plugins.sbt
@@ -1,0 +1,1 @@
+addSbtPlugin("pl.project13.scala" % "sbt-jmh" % "0.3.6")
