--- benchmarking/project/plugins.sbt
@@ -1,1 +1,0 @@
-addSbtPlugin("pl.project13.scala" % "sbt-jmh" % "0.3.6")
