+++ build.sbt
@@ -164,7 +164,7 @@ outputStrategy := Some(StdoutOutput)
 lazy val bifrost = Project(id = "bifrost", base = file("."))
   .settings(commonSettings: _*)
 
-lazy val benchmarking = Project(id = "benchmarking", base = file("benchmarking"))
+lazy val benchmarking = Project(id = "benchmark", base = file("benchmark"))
   .settings(commonSettings: _*)
-  .dependsOn(bifrost)
+  .dependsOn(bifrost % "compile->compile;test->test")
   .disablePlugins(sbtassembly.AssemblyPlugin)
